Convert to bit_order tests
